Problem: Can't get a straight answer "Build what?"
Solution: Add a function to find this straight answer per arch

Taking in a list of packagenames, slowly, very slowly, with lots of
repeated steps, go over every package and dependency to find what's
necessary. Prioritize the bootstrap list, because that is kind of
important.

Currently ignores the bootstrap list for the host arch. Is that
a problem? We will see when we slowly get to the next step.

Would be nice to make this last function lazier, but that might take
unraveling the relatively nested code here.