Problem: XBPS staging breaks remote building
Solution: As a smart man once said, not every problem needs nomad.

In this case, go for a local executor to run builds instead of exporting
to nomad. This works so long as the semantics are the same, and makes
for an excellent case study in using the modularity of this approach.

Who knows, maybe I'll go for lambdacd or some other CI/CD solution to
actually run the builds later. Just for fun, you know?

This fixes the problem we have with xbps and staging right now by
ignoring it. The problem itself has to do with xbps not allowing us to
check the local stagedata repository if we use a remote repository. So,
we need to build from a local repo. How do you do that except with
files and shell building locally?

There are answers (including various FUSE filesystems) but they are all
harder. This is easier.